Add small visual changes
