+++ app/src/main/res/drawable/dradient.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <gradient
+        android:angle="225"
+        android:centerColor="@color/colorWhite"
+        android:endColor="@color/colorPrimary"
+        android:startColor="@color/colorPrimary" />
+</shape>
