+++ app/src/main/res/layout/activity_main.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/activity_main"
     android:layout_width="match_parent"
@@ -12,11 +13,12 @@
         android:layout_height="wrap_content"
         android:background="@color/colorPrimary" />
 
-    <com.creageek.unmaterialtabs.tabs.RoundTabView
+    <com.ruslankishai.unmaterialtab.tabs.RoundTabView
         android:id="@+id/round_tab_view"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="@color/colorPrimary" />
+        android:background="@color/colorPrimary"
+        app:accent="@color/colorAccent" />
 
     <android.support.v4.view.ViewPager
         android:id="@+id/view_pager"
