+++ unmaterialtab/src/main/res/values/attrs.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <declare-styleable name="RoundTabView">
+        <attr name="accent" format="color" />
+    </declare-styleable>
+
+</resources>
