+++ unmaterialtab/src/main/res/values/colors.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <color name="colorPrimary">#37BBAD</color>
-    <color name="colorPrimaryDark">#464646</color>
+    <color name="colorPrimaryDark">#37BBAD</color>
     <color name="colorAccent">#FFFFFF</color>
     <color name="colorWhite">#FFFFFF</color>
 </resources>
